Moved CalendarEventFlexibleAdapter to upper package since it's used by more than just the Month fragment
