+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/calendar/CalendarMonthFragment.kt
@@ -43,8 +43,8 @@ import org.pattonvillecs.pattonvilleapp.SpotlightUtils
 import org.pattonvillecs.pattonvilleapp.calendar.CalendarMonthFragment.Companion.TAG
 import org.pattonvillecs.pattonvilleapp.model.calendar.CalendarRepository
 import org.pattonvillecs.pattonvilleapp.preferences.PreferenceUtils
+import org.pattonvillecs.pattonvilleapp.ui.calendar.CalendarEventFlexibleAdapter
 import org.pattonvillecs.pattonvilleapp.ui.calendar.PinnableCalendarEventItem
-import org.pattonvillecs.pattonvilleapp.ui.calendar.month.CalendarEventFlexibleAdapter
 import org.pattonvillecs.pattonvilleapp.ui.calendar.month.CalendarMonthFragmentViewModel
 import org.pattonvillecs.pattonvilleapp.ui.calendar.month.SingleDayEventFlexibleViewModel
 import org.threeten.bp.DateTimeException
